refactor: use `@apify/tsconfig` with enabled `esModulesInterop`
TS config has slightly changed, the main difference being we use `esModulesInterop`,
and that we no longer emit types to separate folder - they are in the same build folder
next to the source files, together with declaration source maps (`.d.ts.map` files).

Playwright types are no longer inlined, instead the same approach as with puppeteer is
taken (dynamically added `ts-ignore` comments).

As we support only node 15+, the TS `target` and `lib` options have been updated to match
this version.

Keeping the `build` folder for now, I'd switch to `dist` so we are aligned with other
repositories, but that might be actual BC, while this PR should be backwards compatible.