refactor: make the last resort timeout much larger
The `handleRequestTimeout` times twice is definitely not enough for this as there
are things like hooks done inside and those can take a lot on their own. Saw it
often in real life testing. Putting there 10min, as that sounds more like a stuck
run than just twice the request function timeout (which is often just 1-2 minutes).

Related: #1234