fix: rename default storage folder to just `storage`
Given we use the memory storage also in the Actor SDK, we don't want to have a crawlee specific
name for the default storage folder, as it feels confusing to have `crawlee_storage` folder when
you don't interact with `crawlee` anyhow.

This change should not be breaking, if a folder called `crawlee_storage` exists, we still use it.
If not, we use just `storage`.