fix: try to detect stuck request queue and fix its state
When we were facing the RQ stuck issues, the solution was to automatically unstuck it
in case we detect nothing was happening for ~5 minutes and there are some requests in
progress.

This PR removes the condition for in-progress requests, as the queue could get stuck
without any in-progress requests too apparently.

https://apifier.slack.com/archives/C0L33UM7Z/p1679262267148309