feat: allow running single crawler instance multiple times
Often users are trying to reuse single crawler instance, which wasn't possible due to
internal state not being reset in the `run` method completely. One example of this was
the use inside express request handler.

This PR adds support for doing this, by resetting the crawler stats and default request
queue. If the crawler is already running, an error is thrown, suggesting the use of
`crawler.addRequests()` instead.

Related: #765