feat: allow running single crawler instance multiple times (#1844)
Often users are trying to reuse a single crawler instance, which wasn't
possible due to the internal state not being reset in the `run` method
completely. One example of this was the use inside express request
handler.

This PR adds support for doing so, by resetting the crawler stats and
default request queue. If the crawler is already running, an error is
thrown, suggesting the use of `crawler.addRequests()` instead.

Related: #765