feat: accessing crawler state, key-value store and named datasets via crawling context (#2283)
## New/updated `CrawlingContext` methods

- `getKeyValueStore` - replacement for direct `KeyValueStore.open` calls
- `addRequests`, `useState` - basically shortcuts to the respective
methods on the `crawler`
- `pushData` and `getDataset` - there's a new parameter that allows
changing the target dataset

These changes should help the users avoid touching global `Dataset` and
`KeyValueStore` instances and the `crawler` field. Not touching global
crawlee/apify-sdk stuff should become **the way** to write request
handlers, piece by piece.

## `RestrictedCrawlingContext` interface

A subset of the `CrawlingContext` interface that should be enough for
the vast majority of use cases. This is not used anywhere yet. If a user
sticks to using this, the side-effects of the request handler should be
well contained - we can for example run a request handler with a mock
implementation of `RestrictedCrawlingContext` and inspect the "output"
that the handler produced. I intend to use this for a more general
adaptive crawler in the near future.

The restriction should also make it possible to run the request handlers
repeatedly without corrupting the state/dataset/etc. This property can
be useful for example in developer tools, for implementation of new ways
of crawling (beyond adaptive :slightly_smiling_face:), or for making the
existing crawlers more robust in case of errors in the handlers.

I do not intend to remove the original `CrawlingContext` though - if you
find out you need to go "power user mode", you should always be free to
do so. On the other hand, the `RestrictedCrawlingContext` may be too
restricted in some areas. I'm definitely open to judiciously adding back
functionality over time, if needed. Right now, I'd like to keep the
interfaces as simple as possible so that the near-future adaptive
crawler can be implemented with a smaller footprint.

---------

Co-authored-by: Vlad Frangu <kingdgrizzle@gmail.com>