feat: `tieredProxyUrls` for ProxyConfiguration (#2348)
Introduces the `tieredProxyUrls` option for `ProxyConfigration`,
allowing the user to pass proxy URL groups.

The `newProxyInfo` now takes an optional `request` parameter, which can
be used for picking the correct proxy tier. Because of this, the proxy
tiering doesn't work properly in browser crawlers without
`useIncogintoPages` (as the proxy is tied to the launched browser
instance and can be used for multiple requests).