feat: expand #shadow-root elements automatically in `parseWithCheerio` helper (#2396)
Custom HTML elements have their content isolated under a `#shadow-root`
property, this PR handles its expansion by traversing the DOM and
looking for all custom components with a shadow root, inlining its
contents. This way, we can traverse the inside of a custom component via
cheerio, as well as get the text content later on.

The behavior is enabled by default and can be disabled via
`ignoreShadowRoots: true` in the crawler options.