feat(sitemap): Support CDATA in sitemaps (#2424)
During [fixing](https://github.com/topmonks/hlidac-shopu/pull/2109)
Mironet actor for HlidacShopu, I've encountered CDATA in sitemap for the
first time in my life. Not sure how common it is in real world.

https://www.mironet.cz/sm/sitemap_kategorie_p_1.xml.gz

```
<?xml version="1.0" encoding="UTF-8"?>
<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
	<url>
		<loc><![CDATA[https://www.mironet.cz]]></loc>
		<lastmod>2024-04-16</lastmod>
		<changefreq>daily</changefreq>
		<priority>0.9</priority>
	</url>
``` 

This PR is just a quickest implementation I could come up with. Feel
free to make any changes or to disagree that it should be included in
crawlee core :)