fix(core): fire local `SystemInfo` events every second
During local development, we are firing events for the AutoscaledPool about current system resources like memory or CPU. We were firing them once a minute by default, but we remove those snapshots older than 30s, so we never had anything to compare and always used only the very last piece of information.

This PR changes the interval to 1s, aligning this with how the Apify platform fires events.