feat: add `waitForSelector` context helper + `parseWithCheerio` in adaptive crawler
This introduces a new `waitForSelector` context helper to all the crawler types. In browser based crawlers, it will wait for the element, in HTTP crawlers it will check if the selector can be found in the HTML and throw otherwise.

Also, the `parseWithCheerio` is now in adaptive crawler too, and it has a new parameter that allows waiting for a selector via `waitForSelector`.