fix(adaptive-crawler): log only once for the committed request handler execution (#2524)
The adaptive crawler runs some request handlers twice, once via HTTP
client and once via browser. Previously, both of those request handler
executions were producing logs. Now the logging for the HTTP client
execution (which is optional) is proxied and only executed when the
execution gets committed, or an error is thrown. Logging of the
browser-based execution is left intact, since the results from that are
always committed.