fix: fix detection of HTTP site when using the `useState` in adaptive crawler
The crawler state provided in `useState` is now cloned before we run the browser, and the old state is used for the HTTP based rerun, so it can produce the same results. Since we only do this for the reruns, we don't need to care about persistence of the state, we only check for the detection outcome.