docs: fix the Hero section tagline colors (#2757)
The inbuilt Docusaurus CSS minifier doesn't respect the CSS descendant
combinator (likely treats it as an unused style and removes it from the
build). This shows both with the nested CSS and the classic `a b`
syntax. While related CSS issues (e.g.
https://github.com/facebook/docusaurus/issues/9303) are already known to
the Docusaurus team, this one still shows in the latest version
(`docusaurus@3.6.3`).

This PR adds a separate CSS class for the transparent words in the
tagline and explicitly sets this class on the relevant elements