feat: add support for parsing comma-separated list environment variables (#2765)
We're planning to add a new environment variable, `ACTOR_BUILD_TAGS`, to
the Actor runtime, which will contain the list of build tags of the
currently running Actor build. The environment variable will be a
comma-separated list, but I think it would be easier to work with if it
was already parsed to an array in the `Configuration` instance.

This adds support for parsing comma-separated list environment variables
to arrays. In the SDK, I'll then just override the
`COMMA_SEPARATED_LIST_VARS` list to include the environment variable I
want to have parsed.