fix(core): bad signature for Dataset.reduce
BREAKING CHANGE: Dataset.request return type now correctly reflects that `undefined` may be returned
if called on an empty Dataset without the `memo` argument. As calling it without a `memo` argument
already lead to an errorinous type error, users probably already have type casts around it in this
situation, but if not, the new case can be ignored by asserting that there will be a value.

Before:

const res = dataset.reduce(
    async (memo, item, index) => {
        // example impl
        return item;
    }
);

After:

const res = dataset.reduce(
    async (memo, item, index) => {
        // example impl
        return item;
    }
)!;

Note that the original snippet was already rejected by TypeScript for calling reduce with too few
arguments. The workaround to this was to pass `undefined` as the second argument, making
TypeScript infer the type of `res` as `undefined`. This can be "fixed" in one of two ways:
1. In the DatasetReducer, casting the memo argument and the result of reduce to `Data`.
   - In this case no change is neccesary to the user code
2. Calling reduce with the memo argument `undefined as Data`
   - In this case the migration is neccesary