fix(cheerio): don't decode HTML entities in `context.body` (#2838)
`CheerioCrawlingContext.body` now contains a UTF-8 stringified response
body instead of the Cheerio-processed content.

The default Cheerio behaviour introduces a non-injective mapping (both
`"` and `&quot;` are either mapped to `"` or to `&quot;` - based on the
`decodeEntities` parameter). This is likely fine for most use-cases, but
if the content contains non-HTML syntax, it can (irreversibly, because
non-injective) break the contents of the response for the user.

Note that this is somewhere between a fix and a breaking change - while
most(?) devs would assume the new behavior, it does change the content
returned from a public function. I'm fine with this change, but it's
still something to consider.

Closes #2401