chore: fix failing e2e tests (#2854)
Fixes failing e2e tests from e.g.
[here](https://github.com/apify/crawlee/actions/runs/13426528432/job/37510389928)

In `kv-open-return-storage-object`, adds missing `await` (caused the
test to fail with Platform storage backend).

In `puppeteer-store-pagination-jquery-*`, this PR switches the network
wait from the `networkIdle0` logic to `networkIdle2`. This way, the
tests won't fail because of long-running requests. Note that the general
use of `networkIdle` is discouraged, as it brings exactly this kind of
issues. This risk is likely alright in our type of tests (and our
controlled environment), though.

In `cheerio-max-requests` we set the `maxRequestsPerCrawl` to `10`.
Expecting more than `10` requests in the test assert might be valid (due
to concurrent processing), but can cause the tests to fail when we
actually count it right.