feat: Camoufox-based crawler template (#2842)
Adds a Camoufox-based crawler template (`camoufox-ts`).

Compared to the basic `playwright-ts` template, `camoufox-ts` uses the
`camoufox-js` package, which finds the correct latest Camoufox binary in
[GitHub Releases](https://github.com/daijro/camoufox/releases) assets,
downloads it and passes the correct launch options to it.

The `main.ts` script is modified to run the downloaded binary with the
correct `launchOptions`.
Related to #2836