feat: add `respectRobotsTxtFile` crawler option (#2910)
This PR implements automatic skipping of requests based on the
robots.txt file. It works based on a new boolean flag in the crawler
options called `respectRobotsTxtFile`.

Checks are implemented in `BasicCrawler._runTaskFunction()`,
`enqueueLinks` (the context helper), and `BasicCrawler.addRequests()`,
so we ensure disallowed requests are neither processed, nor added to the
queue (this only works with the `enqueueLinks` context helper). The
robots.txt files are cached based on the URL origin, so we should fetch
it only once for each domain (up to 1000 domains, uses LRU cache). We
don't use a proxy to fetch the robots.txt file, which should be fine,
since the file should be accessible by robots in general.

We might want to enable this by default in the next major.