fix(statistics): track actual request.retryCount in Statistics (#2940)
Encountered when debugging run with a migration. A few notes to stats in
general:

1. The statistics.ts is implemented like a generic tracking tool, but it
was ever only used to track request retries and timing, so it would
benefit from being tightly coupled with the Request object. So there is
potential for bigger rewrite, after my change it is only really tracking
time of requests and that doesn't survive migration (not sure if it even
useful for anyone). If we store initial time to Request, we can get rid
of the Job thing.
2. The 
```
"requestsFinished": 97,
"requestsFailed": 11,
"requestsRetries": 44,
```
is really confusing. `requestsRetries` are not total retries but
requests that have at least one retry, those are subset of
`requestsFinished` while `requestsFailed` are completely separate. I
guess renaming would help (not sure how non-breaking we want to be
here), e.g. `requestsFinished` -> `requestsSucceeded` (so it is clear it
doesn't overlap with failed_, `requestsRetries` I would remove
completely and potentially introduce something like
`totalRequestRetries`
3. I still (I already created some issue while back) think it would be
useful to have something like `requestRetryHistogramLive` which would
track requests still in progress. Currently, there is 0 info about how
much you are hitting errors except for the error tracker so
`totalRequestRetries` would also be nice.