fix: Fix useState behavior in adaptive crawler (#2941)
I noticed that the useState() implementation in adaptive playwright
crawler was prone to race conditions. On top of that, during rendering
type detections, the HTTP request handler could receive an older copy of
the state than what the browser handler was working with.

This pull request exposes the crawler state as a mutable value, with all
the associated risks. The copy for the HTTP request handler is made on
demand, during the first useState call in the browser request handler.