fix(social): extract emails from each text node separately (#2952)
This extracts all text nodes from a HTML and calls `emailsFromText` on
each separately. This is to avoid concacenating texts of adjancent
elements, for example:

`<div><a>000 000</a><a>eva@example.com</a></div>`.

`htmlToText` would turn the HTML into `000 000eva@example.com` and we'd
end up with an invalid email `000eva@example.com`.

Co-authored-by: oklinov <oklinov@tutanota.com>