refactor: set storage object on dataset open (#3116)
Following up on [this recent
update](https://github.com/apify/crawlee/pull/2837) where we started
storing the full storageObject when opening a key-value store, this PR
applies the same logic when opening a dataset.

This change prepares the ground for adding a new method:
Dataset.createItemsPublicUrl, which will need access to the dataset's
urlSigningSecretKey.