perf: drop `tsbuildinfo` from published packages (#3243)
Sets `rootDir` (the common root of all build inputs) in Crawlee packages
to `./src`. This causes the `tsbuildinfo` files to get generated in the
package root folder (the one with `package.json`) rather than `/dist`.

The `tsbuildinfo` files are around 175 kB for each package, so this
shrinks all packages by this number of bytes. E.g. `@crawlee/cheerio` is
218 kB of unpacked size as of `3.15.3`, so this is a sizable
improvement.

Closes #3239